# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: isachore API
  description: |
    Chore management for households.

    ## Authenticating

    Two transports, either of which carries a session:

    - the httpOnly `isachore_token` cookie, which a browser sends on its own, and
    - the same opaque token as an `Authorization: Bearer` header, for clients with no cookie jar.

    Two ways to open one, and nothing downstream can tell them apart. `POST /api/v1/auth/login`
    takes an email and password, and answers `two_factor_required` instead of a session when the
    account has TOTP enabled - finish that with `POST /api/v1/auth/verify-2fa`. Where single
    sign-on is configured, `GET /api/v1/auth/oidc/start` is the other way in, and on a server
    with `OIDC_ONLY` set it is the only one: password login answers 403 there.
    `GET /api/v1/auth/methods` says which of the two this server offers, without a session.

    There is no self-registration: administrators create accounts, and the first administrator
    comes from the `init` CLI command.

    ## The CSRF header

    A request authenticated **by cookie** that uses an unsafe method (POST, PATCH, PUT, DELETE)
    must also carry a non-empty `X-CSRF-Token` header, or it is refused with 403 before it
    reaches the route. Any value will do: this is a custom-header defence over `SameSite=Lax`,
    so what matters is that a cross-site form cannot set the header at all.

...
  version: 0.1.0
paths:
...
  /api/v1/completions/filters:
    get:
      tags:
        - completions
      summary: Completion Filters
      description: |-
        The option lists for the History filters: the households the user belongs
        to, and the distinct active members across them (candidate completers).

        Deliberately NOT narrowed by role, unlike the history list right below it. This
        endpoint also feeds the filter bars on Home and Unscheduled (see
        `frontend/src/components/chores/useFilterOptions.ts`), which every role uses, so
        restricting it to deputies would empty the household and member pickers there. Note
        that makes it a superset of what History can show a helper (their own closures only),
        which is why History hides its filter bar entirely rather than narrowing this. It
        exposes household names and member names, both of which every member already sees on
        Home.
      operationId: completion_filters_api_v1_completions_filters_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryFilterOptions'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
      security:
        - sessionCookie: []
        - bearerToken: []
  /api/v1/completions:
    get:
      tags:
        - completions
      summary: List Completions
      description: |-
        Chore history across every active household the caller belongs to, narrowed per
        household by their role: where they are at least a deputy they see everybody's closures,
        where they are only a helper they see their own and nothing else. Reads the `done`
        occurrences of the merged occurrences table, which is both real completions and skips:
        each entry carries `skipped` so the list can tell them apart, and `outcome` narrows to
        one kind. Optional user_id / household_id narrow the list too; a household they do not
        belong to, or a stranger's id in a household they are only a helper in, yields an empty
        page rather than a 403, since the list spans every household at once. History of
        soft-deleted chores is kept (the title is snapshotted for exactly this and the chore row
        still resolves the join).
      operationId: list_completions_api_v1_completions_get
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Page Size
        - name: sort_by
          in: query
          required: false
          schema:
            enum:
              - created_at
              - title
            type: string
            default: created_at
            title: Sort By
        - name: sort_dir
          in: query
          required: false
          schema:
            enum:
              - asc
              - desc
            type: string
            default: desc
            title: Sort Dir
        - name: user_id
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                minimum: 1
              - type: 'null'
            title: User Id
        - name: household_id
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                minimum: 1
              - type: 'null'
            title: Household Id
        - name: outcome
          in: query
          required: false
          schema:
            anyOf:
              - enum:
                  - completed
                  - skipped
                type: string
              - type: 'null'
            title: Outcome
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_HistoryEntryRead_'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/completions/{completion_id}:
    delete:
      tags:
        - completions
      summary: Undo Completion
      description: |-
        Undo a closure, completion or skip alike (identified by its done-occurrence id).
        Either the user it is recorded against (completed_by) or an organiser of that household
        may undo it. Both halves matter: the occurrence stores no separate record of who
        submitted it, so somebody who completes a chore on another member's behalf cannot undo
        it themselves, and before the organiser half existed a helper's closure could be undone
        by nobody at all.
        Undoing the chore's latest closure (the most recently *closed* one, see below)
        reopens that occurrence (deleting the successor open occurrence first) so the chore is
        available again with its original assignee; undoing an older one just removes that
        history row. A skip counts as a closure for "latest" here, because it really did close
        its slot and its successor really is the open row. Note that reopening rolls a rotating
        chore's turn back, so an organiser undoing a housemate's closure can move work between
        people - deliberately, since fixing a mis-skip is the point.
      operationId: undo_completion_api_v1_completions__completion_id__delete
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: completion_id
          in: path
          required: true
          schema:
            type: integer
            title: Completion Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: A member of the household, but the role held there does not reach what this route requires.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No closure with this id in any household you belong to.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ChoreHouseholdRead:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        timezone:
          type: string
          title: Timezone
      type: object
      required:
        - id
        - name
        - timezone
      title: ChoreHouseholdRead
      description: |-
        The household a chore belongs to, embedded in both chore reads (it sits on
        ChoreReadBase) so the list column and the read-only edit header can show its name
        without a second request.

        Also embedded in `DueChoreRead`, `UnscheduledChoreRead`, `HistoryEntryRead` and the
        History filter options, which is what makes `timezone` here reach every household-scoped
        surface at once.
...
    ErrorDetail:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
        - detail
      title: ErrorDetail
      description: |-
        The body FastAPI sends for a raised `HTTPException`.

        Declared so a refusal can be documented as carrying its explanation rather than as
        bodyless, which is what a route with only a `description` claims. Every hand-raised
        HTTPException in this app answers in this shape; only the routes that declare it say
        so, and wiring that up everywhere is the job of the security-scheme step in README's
        todo rather than of any single endpoint.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    HistoryEntryRead:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        scheduled_for:
          type: string
          format: date-time
          title: Scheduled For
        completed_at:
          type: string
          format: date-time
          title: Completed At
        completed_timezone:
          anyOf:
            - type: string
            - type: 'null'
          title: Completed Timezone
        skipped:
          type: boolean
          title: Skipped
        days_late:
          anyOf:
            - type: integer
            - type: 'null'
          title: Days Late
        completed_by:
          anyOf:
            - $ref: '#/components/schemas/HouseholdMemberRead'
            - type: 'null'
        household:
          $ref: '#/components/schemas/ChoreHouseholdRead'
      type: object
      required:
        - id
        - title
        - scheduled_for
        - completed_at
        - completed_timezone
        - skipped
        - days_late
        - completed_by
        - household
      title: HistoryEntryRead
      description: |-
        One closed-occurrence row for the History view.

        `title` is the snapshot taken at closing time (survives a later rename or a
        soft-deleted chore); `completed_at` is when it was actually checked off and
        `scheduled_for` is the occurrence's due datetime, so `days_late` (>0 late, <=0
        on time/early) is the date-based difference between them. It is None for an
        unscheduled chore, which had no due date to be late against: there the slot merely
        records when the chore became available, so subtracting it would report the gap since
        the last completion as lateness. `completed_by` is
        None when the completer's account has been hard-deleted (users are normally
        soft-deleted, so in practice it stays set). Names only on `completed_by`
        (HouseholdMemberRead is data-minimised: no email).

        `skipped` marks the rows where the occurrence was closed without the work being done.
        They belong in the list (something did happen to that slot, and it can be undone like
        any other) but must be readable as distinct from real completions, which is what this
        flag is for. Their `days_late` is None too: there was a deadline, but no work to be
        punctual about.
    HistoryFilterOptions:
      properties:
        households:
          items:
            $ref: '#/components/schemas/ChoreHouseholdRead'
          type: array
          title: Households
        members:
          items:
            $ref: '#/components/schemas/HouseholdMemberRead'
          type: array
          title: Members
      type: object
      required:
        - households
        - members
      title: HistoryFilterOptions
      description: |-
        The option lists for the History filters: the households the current user
        belongs to, and the distinct active members across those households (the
        people who could appear in the "completed by" column).
...
    HouseholdMemberRead:
      properties:
        id:
          type: integer
          title: Id
        first_name:
          type: string
          title: First Name
        last_name:
          type: string
          title: Last Name
      type: object
      required:
        - id
        - first_name
        - last_name
      title: HouseholdMemberRead
...
    Page_HistoryEntryRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/HistoryEntryRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
        - items
        - total
        - page
        - page_size
      title: Page[HistoryEntryRead]
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    sessionCookie:
      type: apiKey
      description: The httpOnly session cookie, set by `POST /api/v1/auth/login` or by the single sign-on callback. A browser sends it automatically; an unsafe method additionally needs the `X-CSRF-Token` header.
      in: cookie
      name: isachore_token
    bearerToken:
      type: http
      description: 'The same opaque session token as an `Authorization: Bearer` header, for clients with no cookie jar. Exempt from the `X-CSRF-Token` requirement.'
      scheme: bearer
    parkedAdminCookie:
      type: apiKey
      description: An impersonating administrator's own session, parked while they act as somebody else. `POST /api/v1/auth/stop-impersonating` is the only route it *authenticates* - several others read it to attribute an action to the real operator, and logout reads it to end both sessions.
      in: cookie
      name: isachore_admin_token